Added neptune logger functions
added metric logging, actions per episode logging and state logging
