+++ .neptune/async/run__9de19a7c-9f37-4569-a49f-001ab6d2aef4__40696__ecm95jhp/metadata.json
@@ -1,0 +1,10 @@
+{
+  "mode": "async",
+  "containerId": "9de19a7c-9f37-4569-a49f-001ab6d2aef4",
+  "containerType": "run",
+  "structureVersion": 3,
+  "os": "Windows-10-10.0.26100-SP0",
+  "pythonVersion": "3.10.11 (tags/v3.10.11:7d4cc5a, Apr  5 2023, 00:38:17) [MSC v.1929 64 bit (AMD64)]",
+  "neptuneClientVersion": "1.13.0",
+  "createdAt": "2025-01-03T16:47:30.991030+00:00"
+}
