+++ archive/logger.py
@@ -1,5 +1,5 @@
 import tensorflow as tf
-
+import utils
 
 class Logger:
 
@@ -60,6 +60,7 @@ class Logger:
         # car1_x, car1_Vx, car2_Vy = car_state[0], car_state[1], car_state[2]
         # print(f"{car_name}_state: car1_Vx={car1_Vx:.2f}, car2_Vy={car2_Vy:.2f}")
         car1_x, same_action = car_state[0], car_state[1]
+        # utils.logger.neptune_logger.log_state_to_neptune(car_state, car_name)
         print(f"{car_name}_state: same_action={same_action}")
 
     def log_q_values(self, q_values, car_name):
