+++ utils/logger/token.json
@@ -1,3 +1,3 @@
 {
-  "api_token": "your_api_token_here"
+  "api_token": "eyJhcGlfYWRkcmVzcyI6Imh0dHBzOi8vYXBwLm5lcHR1bmUuYWkiLCJhcGlfdXJsIjoiaHR0cHM6Ly9hcHAubmVwdHVuZS5haSIsImFwaV9rZXkiOiIyMWE0NDk3NC04MGU1LTQ2ODctOTk4Ni03NjI1MWQ3YzM2ZWYifQ=="
 }
